Tests/LibWeb/Text/input/Messaging/Messaging-post-channel-over-channel.html

<script src="../include.js"></script>
<script>
    asyncTest(done => {
        let channel = new MessageChannel();

        channel.port1.onmessage = (event) => {
            println("Port1: " + JSON.stringify(event.data));
            if (event.ports.length > 0) {
                event.ports[0].postMessage("Hello from the transferred port");
                return;
            }
            channel.port1.postMessage(event.data);
        };

        channel.port2.onmessage = (event) => {
            println("Port2: " + JSON.stringify(event.data));
            if (event.data === "DONE") {
                done();
            }
        };

        let channel2 = new MessageChannel();

        channel2.port2.onmessage = (event) => {
            println("Port3: " + JSON.stringify(event.data))
        }

        channel.port2.postMessage("Hello");
        channel.port2.postMessage({ foo: channel2.port1 }, { transfer: [channel2.port1] });
        channel.port2.postMessage("DONE");
    });
</script>
